Update xMove-VM.ps1
Added several new features
-Unless specified, the move script will use the datastores that the VM currently resides on.  This includes disks that reside on different datastores
-Unless specified, the move script will use the same VDS switch name and portgroup name.  This works for multiple NICs
-Script will now read in vCenter tags (if any) on the source VC, move the VM, then re-apply the tags on the destination. **Note that the tags and categories will need to be created on the destination VC first

Added logic to find the host with the least memory usage inside a cluster.  Also added logic to prevent memory overcommitment during moves.

Added logic to prevent the moves of VM Hardware Version 4.  As specificed in the script, there appears to be a bug that will allow you to move a v4 VM and it will continue to work.  However any change (powercycle, vmotion, etc) will cause the VM's NICs to drop their assigned portgroup.  The only fix (post move that is) appears to be to power it off so the portgroup disappears then re-set the portgroup.

I've only tested the multi-NIC move with a VDS switch but I wrote the logic for a VSS multi-NIC.  Might need troubleshooting.

I've moved approximately 3000 VMs with the script.  Only major issue/annoyance is that large VMs won't move unless there is a ton of freespace on the same datastore.  E.g. if you have a 4TB VM on a 5TB datastore, the move will fail the free space check because the move spec doesn't know that the datastore isn't changing.  Bummer :(