updated for new microcode
Removed most of Intel Sightings logic as you are either on the affected ucode or not. It's unlikely you are going to install one of the withdrawn patches or BIOS updates rather than the new, fixed ones.

Put signatures and ucode revisions into an array (revisions aren't unique across different models).

Added processor signature and whether the ucode is "known good" (we hope) to the table.